and to decimal

+++ ecommerce/apps/inventory/models.py
@@ -64,7 +64,7 @@ class ProductInventory(models.Model):
 
     quantity = models.IntegerField()
     restock_point = models.PositiveIntegerField(blank=True, null=True)
-    weight = models.FloatField()  # in ounces
+    weight = models.DecimalField(decimal_places=2, max_digits=6)  # in ounces
     price = models.DecimalField(max_digits=10, decimal_places=2)
 
     class Meta:
